fixed the css kinda
