+++ client/src/App.js
@@ -3,13 +3,10 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 
 // import Navs from "./components/Nav"
 import Dashboard1 from "./pages/Dashboard";
-import Account from "./pages/Account";
+import ShowAccount from "./pages/Account";
+import Landing from "./pages/Landing";
+import SignOut from "./pages/SignOut";
 
-//clays stuff
-
-import Dashboard from "./components/Dashboard";
-import Account from "./components/Account";
-import Landing from "./components/Landing";
 
 
 
@@ -20,16 +17,11 @@ function App() {
   return (
     <Router>
       <div>
-      {/* <Dashboard /> */}
+          <Route exact path="/" component={Landing} />
           <Route exact path="/dashboard" component={Dashboard1} />
-          <Route exact path="/account" component={Account} />
+          <Route exact path="/account" component={ShowAccount} />
+          <Route exact path="/signout" component={SignOut} />
 
-
-        hello this is a test
-
-      {/* <Dashboard />
-      <Account /> */}
-      <Landing />
     </div>
   </Router>
