+++ client/src/components/Nav/style.css
@@ -14,6 +14,13 @@
     display:none;
 }
 
+.navbar-brand{
+
+
+text-transform: none;
+}
+
+
 
 
 @media only screen and (max-width: 600px) {
@@ -28,7 +35,7 @@
         left: 15%;
         position: relative;
         float: none;
-        top: 15px;
+        top: 0;
 
     }
 
@@ -39,6 +46,10 @@
         color: white;
         top:8px;
         
+    }
+
+    .navbar>.container, .navbar>.container-fluid{
+        justify-content:flex-start;
     }
