+++ client/src/pages/Account.js
@@ -1,15 +1,17 @@
 import React from "react";
 import Navs from "../components/Nav"
+import Account from "../components/Account"
 
 
-function Account() {
+function ShowAccount() {
 	return (
     	<div>
         <Navs page ="Account"/>
+		<Account/>
 
         	Account page
     	</div>
   	);
 }
 
-export default Account;
+export default ShowAccount;
