+++ client/src/pages/SignOut.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Navs from "../components/Nav"
+
+function SignOut() {
+	return (
+    	<div>
+        <Navs page ="Justo"/>
+            Bye
+    	</div>
+  	);
+}
+
+export default SignOut;
